Update package name to : org.calypsonet.keyple.plugin.coppernic + update maven group to : org.calypsonet.keyple + rename plugin module to 'coppernic-plugin'
