+++ coppernic-plugin/.gitignore
