+++ coppernic-plugin/proguard-rules.pro
