+++ coppernic-plugin/src/main/AndroidManifest.xml
@@ -1,0 +1,2 @@
+<manifest package="org.calypsonet.keyple.plugin.coppernic" />
+
