+++ coppernic-plugin/src/main/kotlin/org/calypsonet/keyple/plugin/coppernic/Cone2PluginFactoryProvider.kt
@@ -1,0 +1,23 @@
+/********************************************************************************
+ * Copyright (c) 2021 Calypso Networks Association https://calypsonet.org/
+ *
+ * See the NOTICE file(s) distributed with this work for additional information regarding copyright
+ * ownership.
+ *
+ * This program and the accompanying materials are made available under the terms of the Eclipse
+ * Public License 2.0 which is available at http://www.eclipse.org/legal/epl-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ ********************************************************************************/
+package org.calypsonet.keyple.plugin.coppernic
+
+import android.content.Context
+
+object Cone2PluginFactoryProvider {
+    suspend fun getFactory(context: Context): Cone2PluginFactory {
+        val pluginFactory = Cone2PluginFactoryAdapter()
+        pluginFactory.init(context)
+
+        return pluginFactory
+    }
+}
