+++ example-app/build.gradle
@@ -7,7 +7,7 @@ android {
     buildToolsVersion "29.0.2"
 
     defaultConfig {
-        applicationId "org.eclipse.keyple.coppernic.ask.example"
+        applicationId "org.calypsonet.keyple.plugin.coppernic.example"
         minSdkVersion 19
         targetSdkVersion 29
         versionCode 1
@@ -56,7 +56,7 @@ dependencies {
     implementation "org.calypsonet.terminal:calypsonet-terminal-calypso-java-api:1.0-SNAPSHOT"
     implementation "org.calypsonet.terminal:calypsonet-terminal-card-java-api:1.0-SNAPSHOT"
     
-    implementation project(path: ':keyple-plugins:coppernic-ask')
+    implementation project(path: ':coppernic-plugin')
 
     implementation "com.jakewharton.timber:timber:$timber_version" //Android
