+++ example-app/src/main/kotlin/org/calypsonet/keyple/plugin/coppernic/example/DemoApplication.kt
@@ -1,0 +1,22 @@
+/********************************************************************************
+ * Copyright (c) 2021 Calypso Networks Association https://calypsonet.org/
+ *
+ * See the NOTICE file(s) distributed with this work for additional information regarding copyright
+ * ownership.
+ *
+ * This program and the accompanying materials are made available under the terms of the Eclipse
+ * Public License 2.0 which is available at http://www.eclipse.org/legal/epl-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ ********************************************************************************/
+package org.calypsonet.keyple.plugin.coppernic.example
+
+import android.app.Application
+import timber.log.Timber
+
+class DemoApplication : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        Timber.plant(Timber.DebugTree())
+    }
+}
