--- example-app/src/main/kotlin/org/eclipse/keyple/coppernic/ask/example/model/ChoiceEventModel.kt
@@ -1,15 +1,0 @@
-/********************************************************************************
- * Copyright (c) 2020 Calypso Networks Association https://www.calypsonet-asso.org/
- *
- * See the NOTICE file(s) distributed with this work for additional information regarding copyright
- * ownership.
- *
- * This program and the accompanying materials are made available under the terms of the Eclipse
- * Public License 2.0 which is available at http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- ********************************************************************************/
-package org.eclipse.keyple.coppernic.ask.example.model
-
-data class ChoiceEventModel(val title: String, val choices: List<String> = arrayListOf(), val callback: (choice: String) -> Unit) :
-    EventModel(TYPE_MULTICHOICE, title)
