--- keyple-plugins/coppernic-ask/src/main/AndroidManifest.xml
@@ -1,2 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.eclipse.keyple.coppernic.ask.plugin" />
