--- keyple-plugins/coppernic-ask/src/main/kotlin/org/eclipse/keyple/coppernic/ask/plugin/ParagonSupportedContactlessProtocols.kt
@@ -1,23 +1,0 @@
-/********************************************************************************
- * Copyright (c) 2020 Calypso Networks Association https://www.calypsonet-asso.org/
- *
- * See the NOTICE file(s) distributed with this work for additional information regarding copyright
- * ownership.
- *
- * This program and the accompanying materials are made available under the terms of the Eclipse
- * Public License 2.0 which is available at http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- ********************************************************************************/
-package org.eclipse.keyple.coppernic.ask.plugin
-
-enum class ParagonSupportedContactlessProtocols {
-    ISO_14443,
-    ISO_14443_A,
-    ISO_14443_B,
-    INNOVATRON_B_PRIME,
-    TICKET_CTS_CTM,
-    MIFARE,
-    FELICA,
-    MV5000
-}
