+++ settings.gradle
@@ -1,3 +1,3 @@
 rootProject.name='KeypleAndroidPluginCoppernic'
 include ':example-app'
-include ':keyple-plugins:coppernic-ask'
+include ':coppernic-plugin'
